Move QueueingContextStore to a separate file under Store folder
